+++ lib/appsignal/capistrano.rb
@@ -1,9 +1,5 @@
 require 'capistrano'
-require 'rails'
-require 'appsignal/version'
-require 'appsignal/config'
-require 'appsignal/transmitter'
-require 'appsignal/marker'
+require 'appsignal'
 
 module Appsignal
   class Capistrano
@@ -25,7 +21,7 @@ module Appsignal
 
             marker = Marker.new(marker_data, ENV['PWD'], rails_env, logger)
             if config.dry_run
-              logger.info "Dry run: Deploy marker not actually sent."
+              logger.info("Dry run: Deploy marker not actually sent.")
             else
               marker.transmit
             end