+++ lib/appsignal/transaction.rb
@@ -24,15 +24,16 @@ module Appsignal
       Appsignal.transactions[Thread.current[:appsignal_transaction_id]]
     end
 
-    attr_reader :id, :events, :process_action_event, :action, :exception, :env,
-      :fullpath, :time
+    attr_reader :request_id, :events, :process_action_event, :action, :exception,
+      :env, :fullpath, :time, :tags
 
-    def initialize(id, env)
-      @id = id
+    def initialize(request_id, env)
+      @request_id = request_id
       @events = []
       @process_action_event = nil
       @exception = nil
       @env = env
+      @tags = {}
     end
 
     def sanitized_environment
@@ -45,6 +46,10 @@ module Appsignal
 
     def request
       ActionDispatch::Request.new(@env)
+    end
+
+    def set_tags(given_tags={})
+      @tags.merge!(given_tags)
     end
 
     def set_process_action_event(event)
@@ -80,6 +85,7 @@ module Appsignal
     def truncate!
       process_action_event.payload.clear
       events.clear
+      tags.clear
       sanitized_environment.clear
       sanitized_session_data.clear
       @env = nil
@@ -103,7 +109,7 @@ module Appsignal
 
     def complete!
       Thread.current[:appsignal_transaction_id] = nil
-      current_transaction = Appsignal.transactions.delete(@id)
+      current_transaction = Appsignal.transactions.delete(@request_id)
       if process_action_event || exception?
         Appsignal.enqueue(current_transaction)
       end
@@ -114,7 +120,19 @@ module Appsignal
     def add_sanitized_context!
       sanitize_environment!
       sanitize_session_data!
+      sanitize_tags!
       @env = nil
+    end
+
+    # Only keep tags if they meet the following criteria:
+    # * Key is a symbol or string with less then 100 chars
+    # * Value is a symbol or string with less then 100 chars
+    # * Value is an integer
+    def sanitize_tags!
+      @tags.keep_if do |k,v|
+        (k.is_a?(Symbol) || k.is_a?(String) && k.length <= 100) &&
+        (((v.is_a?(Symbol) || v.is_a?(String)) && v.length <= 100) || (v.is_a?(Integer)))
+      end
     end
 
     def sanitize_environment!
@@ -125,7 +143,7 @@ module Appsignal
 
     def sanitize_session_data!
       @sanitized_session_data =
-        Appsignal::ParamsSanitizer.sanitize(request.session)
+        Appsignal::ParamsSanitizer.sanitize(request.session.to_hash)
       @fullpath = request.fullpath
     end
   end