+++ spec/appsignal/railtie_spec.rb
@@ -19,8 +19,15 @@ describe Appsignal::Railtie do
     end
 
     it "should have set the appsignal subscriber" do
-      Appsignal.subscriber.
-        should be_a ActiveSupport::Notifications::Fanout::Subscriber
+      if defined? ActiveSupport::Notifications::Fanout::Subscribers::Timed
+        # Rails 4
+        Appsignal.subscriber.
+          should be_a ActiveSupport::Notifications::Fanout::Subscribers::Timed
+      else
+        # Rails 3
+        Appsignal.subscriber.
+          should be_a ActiveSupport::Notifications::Fanout::Subscriber
+      end
     end
 
     it "should have added the listener middleware for exceptions" do
@@ -29,7 +36,7 @@ describe Appsignal::Railtie do
 
     context "non action_controller event" do
       it "should call add_event for non action_controller event" do
-        current = mock(:current)
+        current = double(:current)
         current.should_receive(:add_event)
         Appsignal::Transaction.should_receive(:current).twice.
           and_return(current)
@@ -40,7 +47,7 @@ describe Appsignal::Railtie do
 
     context "action_controller event" do
       it "should call set_process_action_event for action_controller event" do
-        current = mock(:current)
+        current = double(:current)
         current.should_receive(:set_process_action_event)
         current.should_receive(:add_event)
         Appsignal::Transaction.should_receive(:current).exactly(3).times.