+++ spec/appsignal/transaction/transaction_formatter_spec.rb
@@ -13,12 +13,12 @@ describe Appsignal::TransactionFormatter do
     context "with a regular request" do
       let(:transaction) { regular_transaction }
 
-      its(:keys) { should == [:request_id, :log_entry, :failed] }
+      its(:keys) { should =~ [:request_id, :log_entry, :failed] }
       its([:request_id]) { should == '1' }
       its([:log_entry]) { should == {
           :action => "BlogPostsController#show",
           :db_runtime => 500,
-          :duration => 100.0,
+          :duration => be_within(0.01).of(100.0),
           :end => 978364860.1,
           :environment => {},
           :kind => "http_request",
@@ -36,14 +36,20 @@ describe Appsignal::TransactionFormatter do
     context "with an exception request" do
       let(:transaction) { transaction_with_exception }
 
-      its(:keys) { should == [:request_id, :log_entry, :failed, :exception] }
+      its(:keys) { should =~ [:request_id, :log_entry, :failed, :exception] }
       its([:request_id]) { should == '1' }
       its([:failed]) { should be_true }
+
+      context "log_entry content" do
+        subject { formatter.hash[:log_entry] }
+
+        its([:tags]) { should == {'user_id' => 123} }
+      end
 
       context "exception content" do
         subject { formatter.hash[:exception] }
 
-        its(:keys) { should == [:backtrace, :exception, :message] }
+        its(:keys) { should =~ [:backtrace, :exception, :message] }
         its([:backtrace]) { should be_instance_of Array }
         its([:exception]) { should == 'ArgumentError' }
         its([:message]) { should == 'oh no' }
@@ -53,7 +59,7 @@ describe Appsignal::TransactionFormatter do
     context "with a slow request" do
       let(:transaction) { slow_transaction }
 
-      its(:keys) { should == [:request_id, :log_entry, :failed, :events] }
+      its(:keys) { should =~ [:request_id, :log_entry, :failed, :events] }
       its([:request_id]) { should == '1' }
       its([:failed]) { should be_false }
 
@@ -63,7 +69,7 @@ describe Appsignal::TransactionFormatter do
         its(:length) { should == 1 }
         its(:first) { should == {
           :name => "query.mongoid",
-          :duration => 100.0,
+          :duration => be_within(0.01).of(100.0),
           :time => 978364860.0,
           :end => 978364860.1,
           :payload => {