+++ spec/appsignal/transmitter_spec.rb
@@ -10,16 +10,17 @@ describe Appsignal::Transmitter do
   describe "#uri" do
     it "returns the uri" do
       Socket.stub(:gethostname => 'app1.local')
-      subject.uri.should == URI(
-        "http://www.80beans.com/action?api_key=the_api_key&"\
-        "hostname=app1.local&gem_version=#{Appsignal::VERSION}"
-      )
+      uri = subject.uri.to_s
+      uri.should include "http://www.80beans.com/action?"
+      uri.should include "api_key=the_api_key"
+      uri.should include "hostname=app1.local"
+      uri.should include "gem_version=#{Appsignal::VERSION}"
     end
   end
 
   describe "#transmit" do
-    let(:response) { mock(:response, :code => '200') }
-    let(:http_client) { mock(:request, :request => response) }
+    let(:response) { double(:response, :code => '200') }
+    let(:http_client) { double(:request, :request => response) }
     before { instance.stub(:http_client => http_client) }
 
     subject { instance.transmit(:shipit => :payload) }
@@ -29,22 +30,24 @@ describe Appsignal::Transmitter do
 
   describe "#http_post" do
     it "calls Net::HTTP.post_form with the correct params" do
-      post = mock(:post)
+      post = double(:post)
       post.should_receive(:[]=).
         with('Content-Type', 'application/json; charset=UTF-8')
-      post.should_receive(:body=).with("{\"the\":\"payload\"}")
+      post.should_receive(:[]=).
+        with('Content-Encoding', 'gzip')
+      post.should_receive(:body=).
+        with(Zlib::Deflate.deflate("{\"the\":\"payload\"}", Zlib::BEST_SPEED))
       Socket.stub(:gethostname => 'app1.local')
 
       Net::HTTP::Post.should_receive(:new).with(
-        "/action?api_key=the_api_key&hostname=app1.local&"\
-        "gem_version=#{Appsignal::VERSION}"
+        instance.uri.request_uri
       ).and_return(post)
       instance.send(:http_post, :the => :payload)
     end
   end
 
-  describe "ca_file_path" do
-    subject { instance.send(:ca_file_path) }
+  describe ".CA_FILE_PATH" do
+    subject { Appsignal::Transmitter::CA_FILE_PATH }
 
     it { should include('resources/cacert.pem') }
     it("should exist") { File.exists?(subject).should be_true }
@@ -64,7 +67,7 @@ describe Appsignal::Transmitter do
 
       its(:use_ssl?) { should be_true }
       its(:verify_mode) { should == OpenSSL::SSL::VERIFY_PEER }
-      its(:ca_file) { include('resources/cacert.pem') }
+      its(:ca_file) { Appsignal::Transmitter::CA_FILE_PATH }
     end
   end
 end