+++ spec/support/helpers/transaction_helpers.rb
@@ -6,6 +6,7 @@ module TransactionHelpers
 
   def transaction_with_exception
     appsignal_transaction.tap do |o|
+      o.set_tags('user_id' => 123)
       begin
         raise ArgumentError, 'oh no'
       rescue ArgumentError => exception
@@ -26,7 +27,7 @@ module TransactionHelpers
       {
         :process_action_event => notification_event(
           :start => fixed_time,
-          :ending => fixed_time + Appsignal.config[:slow_request_threshold] / 1000.0
+          :ending => fixed_time + Appsignal.config[:slow_request_threshold] / 999.99
         )
       }.merge(args)
     )
@@ -37,7 +38,7 @@ module TransactionHelpers
       {
         :process_action_event => notification_event(
           :start => fixed_time,
-          :ending => fixed_time + Appsignal.config[:slow_request_threshold] / 500.0
+          :ending => fixed_time + Appsignal.config[:slow_request_threshold] / 499.99
         )
       }.merge(args)
     )