+++ CHANGELOG.md
@@ -1,6 +1,29 @@
-# 0.5.2
+# 0.6.3
+* Use a mutex around access to the aggregator
 
-* General improvements to the rails generator
+# 0.6.2
+* Bugfix in backtrace cleaner usage for Rails 4
+
+# 0.6.1
+* Bugfix in Capistrano integration
+
+# 0.6.0
+* Support for Rails 4
+* Data that's posted to AppSignal is now gzipped
+* Add Appsignal.send_exception and Appsignal.listen_for_exception
+* We now us the Rails backtrace cleaner
+
+# 0.5.5
+* Fix minor bug
+
+# 0.5.4
+* Debug option in config to get detailed logging
+
+# 0.5.3
+* Fix minor bug
+
+# 0.5.2
+* General improvements to the Rails generator
 * Log to STDOUT if writing to log/appsignal.log is not possible (Heroku)
 * Handle the last transactions before the rails process shuts down
-* require 'erb' to enable loading of appsignal config
+* Require 'erb' to enable dynamic config